+++ timesketch/tsctl.py
@@ -552,7 +552,17 @@ def sketch_delete(sketch_id):
     sketch = Sketch.query.filter_by(id=sketch_id).first()
     if not sketch:
         print("Sketch does not exist.")
+        return
+    elif sketch.get_status.status == "deleted":
+        print(f"Sketch {sketch_id} is already deleted.")
+        return
     else:
+        sketch_labels = [label.label for label in sketch.labels]
+        # do not delete if Sketch has a lithold label
+        if "lithold" in sketch_labels:
+            print("Sketch has lithold label, skipping")
+            return
+
         print(f"Sketch {sketch_id} Name: ({sketch.name})")
         sketch.delete()
         db_session.delete(sketch)
